refactor: refactor reward distributor creation
